requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12
transformers>=4.38.0
diffusers>=0.27.0
peft>=0.9.0
accelerate>=0.27.0
bitsandbytes>=0.42.0
optuna>=3.6.0
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.2.0
opencv-python-headless>=4.9.0
torchmetrics>=1.3.0
grad-cam>=1.5.0
shap>=0.44.0
seaborn>=0.13.0
matplotlib>=3.8.0
